requirements.txt
openai
python-dotenv
gradio
pypdf
pydantic
